requirements.txt
pandas>=2.2
spotipy>=2.23.0
python-dotenv>=1.0.1